+++ spec/rubocop/cli_spec.rb
@@ -254,17 +254,18 @@ describe RuboCop::CLI, :isolated_environment do
       it 'does not trigger UnneededDisable due to lines moving around' do
         src = ['a = 1 # rubocop:disable Lint/UselessAssignment']
         create_file('example.rb', src)
-        create_file('.rubocop.yml', ['Style/Encoding:',
-                                     '  Enabled: true'])
+        create_file('.rubocop.yml', ['Style/FrozenStringLiteralComment:',
+                                     '  Enabled: true',
+                                     '  EnforcedStyle: always'])
         expect(cli.run(['--format', 'offenses', '-a', 'example.rb'])).to eq(0)
         expect($stdout.string).to eq(['',
-                                      '1  Style/Encoding',
+                                      '1  Style/FrozenStringLiteralComment',
                                       '--',
                                       '1  Total',
                                       '',
                                       ''].join("\n"))
         expect(IO.read('example.rb'))
-          .to eq(['# encoding: utf-8',
+          .to eq(['# frozen_string_literal: true',
                   'a = 1 # rubocop:disable Lint/UselessAssignment',
                   ''].join("\n"))
       end